refactor: move error_helpers.py to services/ for proper layering
Move error_helpers.py from exceptions/ to services/ to fix reverse
dependency issue for modularization. This ensures exceptions/ is a
pure foundation layer without runtime dependencies on services/.

Changes:
- Move src/vibe3/exceptions/error_helpers.py → services/error_helpers.py
- Update all 5 call sites to import from services.error_helpers
- Remove re-export in exceptions/__init__.py
- Simplify error_helpers.py imports (no TYPE_CHECKING guard needed)
- Update test fixture to cover services/error_helpers.py
- Update test to check for imports from services.error_helpers
- Update documentation import example
- Remove "Module Placement Trade-off" section (trade-off resolved)

Layer direction now correct:
  exceptions/ (pure definitions) → services/ (business logic)

All tests pass: 7 architecture boundary tests + mypy --strict
