+++ src/vibe3/execution/auto_scene_recovery.py
@@ -105,7 +105,7 @@ class AutoSceneRecoveryService:
         error_msg: str,
     ) -> ExecutionLaunchResult | None:
         from vibe3.exceptions.error_codes import E_EXEC_AUTO_SCENE_RESET
-        from vibe3.exceptions.error_helpers import record_error
+        from vibe3.services.error_helpers import record_error
         from vibe3.services.flow_cleanup_service import (
             FlowCleanupService,
             LiveSessionsDetectedError,
