+++ src/vibe3/execution/codeagent_runner.py
@@ -422,7 +422,7 @@ class CodeagentExecutionService:
                 classify_error_hybrid,
                 get_error_handling_contract,
             )
-            from vibe3.exceptions.error_helpers import record_error
+            from vibe3.services.error_helpers import record_error
 
             # Classify error and record to SQLite for threshold tracking.
             # FailedGate.check() reads SQLite error_log on next heartbeat tick.
